--- src/common/Tiles/MovieTile/TilesContainer/MoviesTitle/index.js
@@ -1,5 +1,0 @@
-import styled from "styled-components";
-
-export const TilesHeader = styled.h1`
-    font-size: 36px;
-`;
