--- src/common/Tiles/MovieTile/TilesContainer/index.js
@@ -1,8 +1,0 @@
-import styled from "styled-components";
-
-export const TilesContainer = styled.article`
-    max-width: 1368px;
-    height: 1391px;
-    margin: auto;
-    margin-top: 56px;
-`;
